extraction assertion to use it outside of pytest
