+++ tests/consensus_tests/downscale_cluster.py
@@ -2,7 +2,7 @@ import argparse
 import requests
 import time
 
-from utils import assert_http_ok
+from assertions import assert_http_ok
 
 parser = argparse.ArgumentParser("Move all shards to first node and detach the last one")
 parser.add_argument("collection_name")
