+++ tests/consensus_tests/test_collection_created_before.py
@@ -1,6 +1,7 @@
 import pathlib
 
 from .utils import *
+from .assertions import assert_http_ok
 
 N_PEERS = 5
 N_SHARDS = 6
